nv-repos/Azur/usecase_beam/Docker/requirements.txt
azure-batch==11.0.0
azure-storage-blob==12.8.1
apache_beam
pandas
pymysql